backend/requirements.txt
fastapi==0.68.1
uvicorn==0.15.0
python-dotenv==0.19.0
firebase-admin==5.0.3
openai==0.27.0
pydantic==1.8.2
python-jose==3.3.0
passlib==1.7.4
python-multipart==0.0.5
requests==2.26.0
